[FIX] account_asset_management: Avoid to skip tests when no CoA found
Since odoo/odoo@d0342c8, default existing company is not getting a CoA
automatically, provoking that these tests are skipped in this case.

We provide a default CoA between existing in that case, and for that,
we have to populate a Odoo environment before calling super for
gathering the needed information, as both the initialization and the
skip is done in super method.